+++ libs/ionic-react/src/utils/versions.ts
@@ -4,6 +4,6 @@ export const ionicReactVersion = '5.0.7';
 export const ionicReactRouterVersion = '5.0.7';
 export const ioniconsVersion = '5.0.1';
 
-export const testingLibraryJestDomVersion = '4.2.4';
+export const testingLibraryJestDomVersion = '5.5.0';
 export const testingLibraryUserEventVersion = '8.1.3';
 export const testingLibraryCypressVersion = '5.3.0';